Patch 1.1.0
Modificato la funzione riguardante la parte decimale  dato che il
riultato è dichiarato come int ma dovrebbe essere float

+++ calcolatricefunzioni.c
@@ -20,7 +20,7 @@ int divisioneINTERA(int a, int b){
 
 float divisioneDECIMALE(float a, float b){
     
-    int risultato;
+    float risultato;
     
     risultato = a/b;
